bot/pyproject.toml
[tool.poetry]
name = "open-interpreter-slack-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
open-interpreter = "^0.1.9"
slack-bolt = "^1.18.0"
Flask = "^2.3.3"
google-cloud-storage = "^2.11.0"
gunicorn = "^21.2.0"
python-json-logger = "^2.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[flake8]
max-line-length = 120
